ceph: Make ceph_calc_file_object_mapping() return size as size_t
Make ceph_calc_file_object_mapping() return the size as a size_t.

cc: Viacheslav Dubeyko <slava@dubeyko.com>
cc: Alex Markuze <amarkuze@redhat.com>
cc: Ilya Dryomov <idryomov@gmail.com>
cc: ceph-devel@vger.kernel.org
cc: linux-fsdevel@vger.kernel.org

+++ fs/ceph/ioctl.c
@@ -192,7 +192,7 @@ static long ceph_ioctl_get_dataloc(struct file *file, void __user *arg)
 		&ceph_sb_to_fs_client(inode->i_sb)->client->osdc;
 	struct ceph_object_locator oloc;
 	CEPH_DEFINE_OID_ONSTACK(oid);
-	u32 xlen;
+	size_t xlen;
 	u64 tmp;
 	struct ceph_pg pgid;
 	int r;
